refactor: split the inline create-dialog flow out of InstallationFormFragment
InstallationFormFragment (877 lines) mixed five concerns: form field
binding, photo handling, save/validation, catalog dropdown wiring, and
the "create Headquarter/Vehicle/Accessory inline" dialog. graphify
flagged this fragment's community as the lowest-cohesion one in the
project, specifically clustering CreateType/ACCESSORY/HEADQUARTER/
VEHICLE together.

Extracted that dialog flow into InstallationCreateController, mirroring
the existing PhotoUploadController(fragment) pattern already used for
photos. It owns the dialog, its validation, and the three
createEvents subscriptions; the fragment only supplies callbacks for
the two side effects it actually needs (applying a newly created
headquarter/vehicle to its own fields).
